+++ src/contracts/interfaces/IERC20Minimal.sol
