+++ src/contracts/interfaces/pool/IUniswapV3PoolDerivedState.sol
