+++ src/contracts/libraries/SafeCast.sol
