+++ test/ConfigurableCorePool.test.ts
@@ -13,11 +13,11 @@ import { TickMath } from "../src/util/TickMath";
 import JSBI from "jsbi";
 import { ZERO } from "../src/enum/InternalConstants";
 import { Transition } from "../src/interface/Transition";
-import { getDate, getTomorrow, format } from "./DateUtils";
-import { EventDBManager } from "./EventDBManager";
-import { EventType } from "./EventType";
-import { LiquidityEvent } from "./LiquidityEvent";
-import { SwapEvent } from "./SwapEvent";
+import { getDate, getTomorrow, format } from "../src/util/DateUtils";
+import { EventDBManager } from "../src/manager/EventDBManager";
+import { EventType } from "../src/enum/EventType";
+import { LiquidityEvent } from "../src/entity/LiquidityEvent";
+import { SwapEvent } from "../src/entity/SwapEvent";
 import { printParams } from "../src/util/Serializer";
 import { SimulatorConsoleVisitor } from "../src/manager/SimulatorConsoleVisitor";
 import { SimulatorPersistenceVisitor } from "../src/manager/SimulatorPersistenceVisitor";
