--- test/EventType.ts
@@ -1,5 +1,0 @@
-export enum EventType {
-  MINT = 1,
-  BURN = 2,
-  SWAP = 3,
-}
